azfunc/requirements.txt
azure-functions
pandas
numpy
git+git://github.com/loyning/python-24so#egg=tfsoffice
sqlalchemy
pyodbc